requirements.txt
# Tech Digest Python Dependencies
# Install with: pip install -r requirements.txt

# RSS parsing (optional, will fallback to regex if not available)
feedparser>=6.0.0

# JSON Schema validation (optional)
jsonschema>=4.0.0

# Postgres storage (required for DB pipeline mode)
psycopg2-binary>=2.9.0

# Load .env file automatically (required)
python-dotenv>=1.0.0

# Standard library modules used:
# - json
# - urllib.request
# - urllib.parse
# - urllib.error
# - concurrent.futures
# - datetime
# - pathlib
# - argparse
# - logging
# - tempfile
# - difflib
# - re
# - time
# - sys
# - os

# Note: All scripts are designed to work with Python 3.8+ standard library only.
# The optional dependencies above enhance functionality but are not required.